pyproject.toml
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.

[tool.black]
line-length = 80
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = 'halsey'
version = "2.0.0"
description = ''
authors = ['Jared Coughlin']

[tool.poetry.dependencies]
python = '^3.6'
numpy = "^1.17"
gym = "^0.15.4"
atari_py = "^0.2.6"
tensorflow = "~2.1"
sphinx = "^2.3.1"
sphinx_rtd_theme = "^0.4.3"
gin-config = "^0.3.0"
rich = "^0.8.6"
h5py = "^2.10.0"
tensorflow-estimator = "2.1.0"
pyyaml = "^5.3.1"
scipy = "^1.4.1"

[tool.poetry.dev-dependencies]
ipython = "^7.9"
black = "^18.9b0"
flake8 = "^3.6"

[tool.poetry.scripts]
halsey = 'halsey.main.main:run'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
